+++ src/app/pages/department.closure/components/IDepartmentClosureService.ts
@@ -5,7 +5,7 @@ import { IServiceInretface, ResponseModel } from '../../../shared';
 
 export interface IDepartmentClosureService extends IServiceInretface<DepartmentClosureModel> {
 
-    GetAllByIncident(incidentId: number): Observable<ResponseModel<DepartmentClosureModel>>;
+    GetAllByIncident(incidentId: number,departmentId:number): Observable<ResponseModel<DepartmentClosureModel>>;
 
     getAllbyIncidentandDepartment(incidentId, departmentId): Observable<ResponseModel<DepartmentClosureModel>>;
 }
