+++ src/app/pages/department.closure/components/index.ts
@@ -1,2 +1,3 @@
 export * from './department.closure.service';
 export * from './department.closure.model';
+export * from './IDepartmentClosureService';
