+++ src/app/pages/department.closure/index.ts
@@ -1,1 +1,2 @@
 export * from './components';
+export * from './department.closure.module';
