+++ src/app/pages/department.closure/styles/department.closure.style.scss
