--- src/app/pages/departmentclosure/components/index.ts
@@ -1,3 +1,0 @@
-export * from './department.closure.service';
-export * from './department.closure.model';
-export * from './IDepartmentClosureService';
