--- src/app/pages/departmentclosure/index.ts
@@ -1,2 +1,0 @@
-export * from './components';
-export * from './department.closure.module';
