+++ src/app/pages/pages.routing.ts
@@ -24,7 +24,7 @@ const pageRoutes: Routes = [
             { path: 'incident', loadChildren: './incident/incident.module#IncidentModule' },
             { path: 'masterdataupload', loadChildren: './masterdata.upload/masterdata.upload.module#MasterDataUploadModule' },
             { path: 'notifypeople', loadChildren: './notifypeople/notifypeople.module#NotifyPeopleModule' },
-            { path: 'departmentclosure', loadChildren: './departmentclosure/department.closure.module#DepartmentClosureModule' },
+            { path: 'departmentclosure', loadChildren: './department.closure/department.closure.module#DepartmentClosureModule' },
             { path: 'emergencyclosure', loadChildren: './emergency.closure/emergency.closure.module#EmergencyClosureModule' },
             { path: 'notifypeople', loadChildren: './notifypeople/notifypeople.module#NotifyPeopleModule' }
         ]
